LayoutTests/canvas/philip/tests/2d.path.moveTo.nonfinite-expected.txt

< [index] >
2d.path.moveTo.nonfinite
moveTo() with Infinity/NaN is ignored
References: 2d.nonfinite
Actual output:
Expected output:

Passed
